Handle odd list-tag-ID in Type-Chunk.
